load-generator/requirements.txt
kafka-python==2.0.2
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
prometheus-client==0.19.0